+++ lib/extensions/reset_password/ecto/context.ex
@@ -1,10 +1,10 @@
 defmodule PowResetPassword.Ecto.Context do
   @moduledoc false
-  alias Pow.{Config, Ecto.Context}
+  alias Pow.{Config, Ecto.Context, Operations}
   alias PowResetPassword.Ecto.Schema
 
   @spec get_by_email(binary(), Config.t()) :: Context.user() | nil
-  def get_by_email(email, config), do: Context.get_by([email: email], config)
+  def get_by_email(email, config), do: Operations.get_by([email: email], config)
 
   @spec update_password(Context.user(), map(), Config.t()) :: {:ok, Context.user()} | {:error, Context.changeset()}
   def update_password(user, params, config) do